# tools-workspace/contracts/pdf/openapi.yaml
...
openapi: 3.0.3
info:
  title: EasyToolHub Tool API — PDF domain
  version: 0.1.0
  description: Client-upload PDF microservice for EasyToolHub (temp processing, no permanent storage).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/pdf/batch:
    post:
      summary: Submit an async batch job (ZIP of PDFs or multiple PDFs)
      description: >
        Returns a job id immediately. Poll GET /api/v1/pdf/jobs/{id} for progress,
        then GET /api/v1/pdf/jobs/{id}/download for the result ZIP.
        Operations: remove-metadata, remove-annotations, remove-hidden, rotate,
        watermark, page-numbers, flatten-form, repair.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary, description: ZIP of PDFs or a single PDF }
                files:
                  type: array
                  items: { type: string, format: binary }
                  description: Multiple PDFs and/or ZIPs
                operation:
                  type: string
                  default: remove-metadata
                degrees: { type: integer, default: 90 }
                watermarkText: { type: string, default: CONFIDENTIAL }
                opacity: { type: number, default: 0.3 }
                startAt: { type: integer, default: 1 }
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
  /api/v1/pdf/jobs/{id}:
    get:
      summary: Poll batch job status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
        "404":
          description: Job not found or expired
  /api/v1/pdf/jobs/{id}/download:
    get:
      summary: Download completed batch result ZIP
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Result ZIP of processed PDFs
          content:
            application/zip:
              schema: { type: string, format: binary }
        "422":
          description: Job not ready
        "404":
          description: Job not found or expired
...
  /api/v1/pdf/office-to-pdf-async:
    post:
      summary: Convert Office document to PDF (async)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
...
  /api/v1/pdf/url-to-pdf-async:
    post:
      summary: Render a public URL to PDF asynchronously
      parameters:
        - in: query
          name: url
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfJobStatus'
...
components:
  schemas:
    PdfJobStatus:
      type: object
      properties:
        jobId: { type: string }
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, EXPIRED]
        operation: { type: string }
        total: { type: integer }
        completed: { type: integer }
        failed: { type: integer }
        progress: { type: number, minimum: 0, maximum: 1 }
        message: { type: string }
        downloadReady: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }